requirements.txt
amqp==5.2.0
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backoff==1.10.0
billiard==4.2.1
black==23.10.1
boto3==1.34.100
botocore==1.34.100
build==0.7.0
celery==5.3.4
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.0.0
decorator==5.1.1
distlib==0.3.9
distro-info==1.0
Django==4.2.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==24.2
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
docutils==0.21.2
drf-extra-fields==3.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.115.7
ffmpeg-python==0.2.0
filelock==3.17.0
fiscalyear==0.4.0
flake8==6.1.0
fsspec==2024.12.0
future==1.0.0
gTTS==2.5.4
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
huggingface-hub==0.28.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.4.5
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.5
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka==1.3.5
kombu==5.4.2
llvmlite==0.44.0
lml==0.1.0
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
monotonic==1.6
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nydus==0.11.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.11.0.86
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pay-ccavenue==0.1.6
pep517==0.12.0
phonenumbers==8.12.43
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.6.0
proglog==0.1.10
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.12.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyexcel-io==0.6.7
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pyflakes==3.1.0
Pygments==2.18.0
pyHanko==0.25.2
pyhanko-certvalidator==0.26.4
PyJWT==2.3.0
pymongo==4.11
# pymssql==2.2.10
pyodbc==5.1.0
pyOpenSSL==24.3.0
pyparsing==3.0.7
pypdf==5.1.0
pytest==7.4.3
pytest-django==4.5.2
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2021.3
PyYAML==6.0.2
qrcode==8.0
redgreenunittest==0.1.1
redis==5.0.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.31.0
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.5.2
scenedetect==0.6.5.2
sentry-sdk==2.1.1
service-identity==24.2.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
starlette==0.45.3
svglib==1.5.1
swagger-spec-validator==3.0.4
sympy==1.13.1
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.1.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.34.0
uWSGI==2.0.28
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
zope.interface==7.2
django-celery-results==2.5.1
